Fixes for the usage of DDP_EMPTY_STRING

+++ lib/stdlib/source/filesystem.c
@@ -36,16 +36,25 @@
 #endif // DDPOS_WINDOWS
 
 ddpbool Existiert_Pfad(ddpstring *Pfad) {
+	if (ddp_string_empty(Pfad)) {
+		return false;
+	}
 	return access(Pfad->str, F_OK) == 0;
 }
 
 ddpbool Erstelle_Ordner(ddpstring *Pfad) {
+	DDP_MIGHT_ERROR;
+
+	if (ddp_string_empty(Pfad)) {
+		ddp_error("Der gegebene Pfad ist leer", false);
+		return false;
+	}
 	// recursively create every directory needed to create the final one
 	char *it = Pfad->str;
 	while ((it = strpbrk(it, PATH_SEPERATOR)) != NULL) {
 		*it = '\0';
 		if (mkdir(Pfad->str) != 0 && errno != EEXIST) {
-			ddp_error("Fehler beim Erstellen des Ordners '"DDP_STRING_FMT"': ", true, Pfad->str);
+			ddp_error("Fehler beim Erstellen des Ordners '" DDP_STRING_FMT "': ", true, Pfad->str);
 			return false;
 		}
 		*it = '/';
@@ -56,13 +65,19 @@ ddpbool Erstelle_Ordner(ddpstring *Pfad) {
 	if (Pfad->str[Pfad->cap - 2] == '/') {
 		return true;
 	} else if (mkdir(Pfad->str) != 0 && errno != EEXIST) {
-		ddp_error("Fehler beim Erstellen des Ordners '"DDP_STRING_FMT"': ", true, Pfad->str);
+		ddp_error("Fehler beim Erstellen des Ordners '" DDP_STRING_FMT "': ", true, Pfad->str);
 		return false;
 	}
 	return true;
 }
 
 ddpbool Ist_Ordner(ddpstring *Pfad) {
+	DDP_MIGHT_ERROR;
+
+	if (ddp_string_empty(Pfad)) {
+		return false;
+	}
+
 	// remove possible trailing seperators
 	char *it = Pfad->str + Pfad->cap - 2; // last character in str
 	while (it >= Pfad->str) {
@@ -75,7 +90,7 @@ ddpbool Ist_Ordner(ddpstring *Pfad) {
 
 	struct stat path_stat;
 	if (stat(Pfad->str, &path_stat) != 0) {
-		ddp_error("Fehler beim Überprüfen des Pfades '"DDP_STRING_FMT"': ", true, Pfad->str);
+		ddp_error("Fehler beim Überprüfen des Pfades '" DDP_STRING_FMT "': ", true, Pfad->str);
 		return false;
 	}
 	return S_ISDIR(path_stat.st_mode);
@@ -106,7 +121,7 @@ static int remove_directory(const char *path) {
 			if (buf) {
 				struct stat statbuf;
 
-				snprintf(buf, len, ""DDP_STRING_FMT"/"DDP_STRING_FMT"", path, p->d_name);
+				snprintf(buf, len, "" DDP_STRING_FMT "/" DDP_STRING_FMT "", path, p->d_name);
 				if (!stat(buf, &statbuf)) {
 					if (S_ISDIR(statbuf.st_mode)) {
 						r2 = remove_directory(buf);
@@ -120,14 +135,14 @@ static int remove_directory(const char *path) {
 		}
 		closedir(d);
 	} else {
-		ddp_error("Fehler beim Öffnen des Ordners '"DDP_STRING_FMT"': ", true, path);
+		ddp_error("Fehler beim Öffnen des Ordners '" DDP_STRING_FMT "': ", true, path);
 		return -1;
 	}
 
 	if (!r) {
 		r = rmdir(path);
 	} else {
-		ddp_error("Fehler beim Löschen des Ordners '"DDP_STRING_FMT"': ", true, path);
+		ddp_error("Fehler beim Löschen des Ordners '" DDP_STRING_FMT "': ", true, path);
 		return -1;
 	}
 
@@ -135,20 +150,34 @@ static int remove_directory(const char *path) {
 }
 
 ddpbool Loesche_Pfad(ddpstring *Pfad) {
+	DDP_MIGHT_ERROR;
+
+	if (ddp_string_empty(Pfad)) {
+		ddp_error("Der gegebene Pfad ist leer", false);
+		return false;
+	}
+
 	if (Ist_Ordner(Pfad)) {
 		return remove_directory(Pfad->str) == 0;
 	}
 	if (unlink(Pfad->str) != 0) {
-		ddp_error("Fehler beim Löschen des Pfades '"DDP_STRING_FMT"': ", true, Pfad->str);
+		ddp_error("Fehler beim Löschen des Pfades '" DDP_STRING_FMT "': ", true, Pfad->str);
 		return false;
 	}
 	return true;
 }
 
 ddpbool Pfad_Verschieben(ddpstring *Pfad, ddpstring *NeuerName) {
+	DDP_MIGHT_ERROR;
+
+	if (ddp_string_empty(Pfad)) {
+		ddp_error("Der gegebene Pfad ist leer", false);
+		return false;
+	}
+
 	struct stat path_stat;
 	if (stat(NeuerName->str, &path_stat) != 0) {
-		ddp_error("Fehler beim Überprüfen des Pfades '"DDP_STRING_FMT"': ", true, NeuerName->str);
+		ddp_error("Fehler beim Überprüfen des Pfades '" DDP_STRING_FMT "': ", true, NeuerName->str);
 		return false;
 	}
 
@@ -168,7 +197,7 @@ ddpbool Pfad_Verschieben(ddpstring *Pfad, ddpstring *NeuerName) {
 		DDP_FREE(char, path_copy);
 	}
 	if (rename(Pfad->str, NeuerName->str) != 0) {
-		ddp_error("Fehler beim Verschieben des Pfades '"DDP_STRING_FMT"' nach '"DDP_STRING_FMT"': ", true, Pfad->str, NeuerName->str);
+		ddp_error("Fehler beim Verschieben des Pfades '" DDP_STRING_FMT "' nach '" DDP_STRING_FMT "': ", true, Pfad->str, NeuerName->str);
 		return false;
 	}
 	return true;
@@ -189,10 +218,12 @@ static void formatDateStr(ddpstring *str, struct tm *time) {
 }
 
 void Zugriff_Datum(ddpstring *ret, ddpstring *Pfad) {
+	DDP_MIGHT_ERROR;
+
 	struct stat st;
 	if (stat(Pfad->str, &st) != 0) {
 		*ret = (ddpstring){0};
-		ddp_error("Fehler beim Überprüfen des Pfades '"DDP_STRING_FMT"': ", true, Pfad->str);
+		ddp_error("Fehler beim Überprüfen des Pfades '" DDP_STRING_FMT "': ", true, Pfad->str);
 		return;
 	}
 	struct tm *tm = localtime(&st.st_atime);
@@ -201,10 +232,12 @@ void Zugriff_Datum(ddpstring *ret, ddpstring *Pfad) {
 }
 
 void AEnderung_Datum(ddpstring *ret, ddpstring *Pfad) {
+	DDP_MIGHT_ERROR;
+
 	struct stat st;
 	if (stat(Pfad->str, &st) != 0) {
 		*ret = (ddpstring){0};
-		ddp_error("Fehler beim Überprüfen des Pfades '"DDP_STRING_FMT"': ", true, Pfad->str);
+		ddp_error("Fehler beim Überprüfen des Pfades '" DDP_STRING_FMT "': ", true, Pfad->str);
 		return;
 	}
 	struct tm *tm = localtime(&st.st_mtime);
@@ -213,10 +246,12 @@ void AEnderung_Datum(ddpstring *ret, ddpstring *Pfad) {
 }
 
 void Status_Datum(ddpstring *ret, ddpstring *Pfad) {
+	DDP_MIGHT_ERROR;
+
 	struct stat st;
 	if (stat(Pfad->str, &st) != 0) {
 		*ret = (ddpstring){0};
-		ddp_error("Fehler beim Überprüfen des Pfades '"DDP_STRING_FMT"': ", true, Pfad->str);
+		ddp_error("Fehler beim Überprüfen des Pfades '" DDP_STRING_FMT "': ", true, Pfad->str);
 		return;
 	}
 	struct tm *tm = localtime(&st.st_ctime);
@@ -225,9 +260,11 @@ void Status_Datum(ddpstring *ret, ddpstring *Pfad) {
 }
 
 ddpint Datei_Groesse(ddpstring *Pfad) {
+	DDP_MIGHT_ERROR;
+
 	struct stat st;
 	if (stat(Pfad->str, &st) != 0) {
-		ddp_error("Fehler beim Überprüfen des Pfades '"DDP_STRING_FMT"': ", true, Pfad->str);
+		ddp_error("Fehler beim Überprüfen des Pfades '" DDP_STRING_FMT "': ", true, Pfad->str);
 		return -1;
 	}
 
@@ -235,9 +272,11 @@ ddpint Datei_Groesse(ddpstring *Pfad) {
 }
 
 ddpint Datei_Modus(ddpstring *Pfad) {
+	DDP_MIGHT_ERROR;
+
 	struct stat st;
 	if (stat(Pfad->str, &st) != 0) {
-		ddp_error("Fehler beim Überprüfen des Pfades '"DDP_STRING_FMT"': ", true, Pfad->str);
+		ddp_error("Fehler beim Überprüfen des Pfades '" DDP_STRING_FMT "': ", true, Pfad->str);
 		return -1;
 	}
 
@@ -246,6 +285,13 @@ ddpint Datei_Modus(ddpstring *Pfad) {
 
 // UNIX: https://stackoverflow.com/questions/2180079/how-can-i-copy-a-file-on-unix-using-c
 ddpbool Datei_Kopieren(ddpstring *Pfad, ddpstring *Kopiepfad) {
+	DDP_MIGHT_ERROR;
+
+	if (ddp_string_empty(Pfad)) {
+		ddp_error("Der gegebene Pfad ist leer", false);
+		return -1;
+	}
+
 #ifdef DDPOS_WINDOWS
 	return (ddpbool)CopyFile(Pfad->str, Kopiepfad->str, false);
 #else  // DDPOW_LINUX
@@ -301,6 +347,8 @@ out_error:
 }
 
 ddpint Lies_Text_Datei(ddpstring *Pfad, ddpstringref ref) {
+	DDP_MIGHT_ERROR;
+
 	FILE *file = fopen(Pfad->str, "r");
 	if (file) {
 		fseek(file, 0, SEEK_END);			  // seek the last byte in the file
@@ -312,25 +360,27 @@ ddpint Lies_Text_Datei(ddpstring *Pfad, ddpstringref ref) {
 		fclose(file);
 		ref->str[ref->cap - 1] = '\0';
 		if (read != string_size - 1) {
-			ddp_error("Fehler beim Lesen der Datei '"DDP_STRING_FMT"': ", true, Pfad->str);
+			ddp_error("Fehler beim Lesen der Datei '" DDP_STRING_FMT "': ", true, Pfad->str);
 		}
 		return (ddpint)read;
 	}
-	ddp_error("Fehler beim Öffnen der Datei '"DDP_STRING_FMT"': ", true, Pfad->str);
+	ddp_error("Fehler beim Öffnen der Datei '" DDP_STRING_FMT "': ", true, Pfad->str);
 	return -1;
 }
 
 ddpint Schreibe_Text_Datei(ddpstring *Pfad, ddpstring *text) {
+	DDP_MIGHT_ERROR;
+
 	FILE *file = fopen(Pfad->str, "w");
 	if (file) {
 		int ret = fprintf(file, text->str);
 		fclose(file);
 		if (ret < 0) {
-			ddp_error("Fehler beim Schreiben der Datei '"DDP_STRING_FMT"': ", true, Pfad->str);
+			ddp_error("Fehler beim Schreiben der Datei '" DDP_STRING_FMT "': ", true, Pfad->str);
 			return ret;
 		}
 		return (ddpint)ret;
 	}
-	ddp_error("Fehler beim Öffnen der Datei '"DDP_STRING_FMT"': ", true, Pfad->str);
+	ddp_error("Fehler beim Öffnen der Datei '" DDP_STRING_FMT "': ", true, Pfad->str);
 	return -1;
 }
